implant-cv/src/requirements.txt
# Ensure compatibility with SageMaker's default environment
numpy<2.0,>=1.21.0
pillow>=9.0.0

# PyTorch & torchvision for model training
torch>=2.0.0
torchvision>=0.15.0

# For dataset annotation and visualization
scikit-learn>=1.0.0
albumentations>=1.3.0
opencv-python>=4.8.0

# Label Studio JSON parsing + S3
boto3>=1.24.0
PyYAML>=6.0
tqdm>=4.64.0

# Optional: for segmentation-style workflows
ultralytics>=8.2.0
supervision>=0.15.0
pycocotools>=2.0.6
shapely>=2.0.0
rasterio>=1.3.0

# Visualization
matplotlib>=3.6.0
seaborn>=0.12.0

# Transformers/utilities (optional for later expansion)
timm>=0.9.0
einops>=0.7.0

# Logging and experiment tracking
wandb>=0.15.0

